% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deng_count_data}
\alias{deng_count_data}
\title{Test dataset}
\format{An object of class \code{matrix} with 18884 rows and 286 columns.}
\usage{
deng_count_data
}
\description{
A numeric matrix with 18884 rows and 315 columns. This matrix can be
used to test the function \code{cell_clustering} provided by the package.
}
\keyword{datasets}
