% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{preprocessing}
\alias{preprocessing}
\alias{log_normalization}
\title{Normalization}
\usage{
preprocessing(x, percent)

log_normalization(x)
}
\arguments{
\item{x}{The expression matrix that needs preprocessing or log-normalization.}

\item{percent}{Genes that are expressed in less than 100*percent\% of the cells are filtered out.}
}
\value{
The preprocessed and log-normalized expression matrix.
}
\description{
Performs preprocessing and log-normalization on the raw count dataset.
}
